refactor: move periodic_sigint to tests/util.py
